# ServerAI/swaggerDefinitions.yaml
swagger: '2.0'
info:
  title: Farming Recommendation API
  description: API for irrigation recommendations and user management
  version: 1.0.0
...
schemes:
  - https
paths:
...
  /forum/newMessage:
    post:
      summary: Add a new forum message
      tags:
        - Forum
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          description: New forum message object
          required: true
          schema:
            $ref: '#/definitions/NewForumMessage'
      responses:
        '201':
          description: Message created successfully
          schema:
            $ref: '#/definitions/ForumMessage'
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error
...
  /forum/messages:
    get:
      summary: Get all forum messages
      tags:
        - Forum
      responses:
        '200':
          description: Forum messages retrieved successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/ForumMessage'
        '400':
          description: Bad Request
        '404':
          description: Forum messages not found
        '500':
          description: Internal Server Error
  /forum/messages/{messageId}/comments:
    get:
      summary: Get all comments for a specific forum message
      tags:
        - Forum
      parameters:
        - in: path
          name: messageId
          required: true
          type: string
          description: ID of the forum message to retrieve comments for
      responses:
        '200':
          description: Comments retrieved successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/ForumComment'
        '400':
          description: Bad Request
        '404':
          description: Comments not found
        '500':
          description: Internal Server Error
definitions:
...
  ForumMessage:
    type: object
    properties:
      _id:
        type: string
        description: The message ID
      userId:
        type: string
        description: ID of the user who posted the message
      content:
        type: string
        description: Content of the message
  NewForumMessage:
    type: object
    properties:
      userId:
        type: string
        description: ID of the user posting the message
      content:
        type: string
        description: Content of the message
  ForumComment:
    type: object
    properties:
      _id:
        type: string
        description: The comment ID
      messageId:
        type: string
        description: ID of the message the comment belongs to
      userId:
        type: string
        description: ID of the user who posted the comment
      content:
        type: string
        description: Content of the comment
...